+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## [Unreleased]
 
 ## [0.0.3] - 2024-09-05
-- Organisation service.
+- Added Organisation service with CRUD operations and `add_org_to_token`. [#7](https://github.com/Swiftner/swiftner_ruby/pull/7)
 
 ## [0.0.2] - 2024-02-13