+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## [Unreleased]
 
 ## [0.0.3] - 2024-09-05
-- Chapter service.
+- Added Chapter service with CRUD operations. [#8](https://github.com/Swiftner/swiftner_ruby/pull/8)
+
  
 ## [0.0.2] - 2024-02-13