+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## [Unreleased]
-
+- Added Chapter service with CRUD operations. [#8](https://github.com/Swiftner/swiftner_ruby/pull/8)
 - Added Organisation service with CRUD operations and `add_org_to_token`. [#7](https://github.com/Swiftner/swiftner_ruby/pull/7)
 
 ## [0.0.2] - 2024-02-13